requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
playwright==1.48.0
twilio==9.3.2
python-dotenv==1.0.1
